# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation for the project",
        "version": "1.0"
    },
    "paths": {
        "/api/autos": {
            "get": {
                "summary": "Get list of autos",
                "operationId": "779873266044adce068514df6f023ec5",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "summary": "Create a new auto",
                "operationId": "3775832f6b13251c20a7e79e3d7b61c9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Modelo",
                                    "Motor",
                                    "ConstructorId",
                                    "EquipoId"
                                ],
                                "properties": {
                                    "Modelo": {
                                        "type": "string",
                                        "example": "Modelo A"
                                    },
                                    "Motor": {
                                        "type": "string",
                                        "example": "Motor A"
                                    },
                                    "ConstructorId": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "EquipoId": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/autos/{id}": {
            "get": {
                "summary": "Get auto by ID",
                "operationId": "1be087abed0ba160d1479ce6f78168b3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "summary": "Update auto by ID",
                "operationId": "7a7f900b210d60721c4cf9b02b94d2b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Modelo",
                                    "Motor",
                                    "ConstructorId",
                                    "EquipoId"
                                ],
                                "properties": {
                                    "Modelo": {
                                        "type": "string",
                                        "example": "Modelo A"
                                    },
                                    "Motor": {
                                        "type": "string",
                                        "example": "Motor A"
                                    },
                                    "ConstructorId": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "EquipoId": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "delete": {
                "summary": "Delete auto by ID",
                "operationId": "ebf42d58970816527f1e8fc42f8c811d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/campeonatos": {
            "get": {
                "summary": "Get list of campeonatos",
                "operationId": "b7ee36ca8c11af3c5847ec45cb6a32ed",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "summary": "Create a new campeonato",
                "operationId": "033bd5c0c9fc9432b9d796394926c50c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Nombre",
                                    "Año"
                                ],
                                "properties": {
                                    "Nombre": {
                                        "type": "string",
                                        "example": "Campeonato A"
                                    },
                                    "Año": {
                                        "type": "integer",
                                        "example": 2023
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/campeonatos/{id}": {
            "get": {
                "summary": "Get campeonato by ID",
                "operationId": "45f94d7e250906803ddc44f89b5cbb47",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "summary": "Update campeonato by ID",
                "operationId": "3b1155d7af60c30851b659d1f3422ed4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Nombre",
                                    "Año"
                                ],
                                "properties": {
                                    "Nombre": {
                                        "type": "string",
                                        "example": "Campeonato A"
                                    },
                                    "Año": {
                                        "type": "integer",
                                        "example": 2023
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "delete": {
                "summary": "Delete campeonato by ID",
                "operationId": "73d2098d04551a458b1db3cc231d349a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
    }
}